+++ app/src/main/java/com/example/spoot_taxi_front/activities/MainActivity.java
@@ -73,23 +73,6 @@ public class MainActivity extends AppCompatActivity {
                 Log.d("FCM_TEST", notificationData);
         }
 
-        FirebaseMessaging.getInstance().getToken()
-                .addOnCompleteListener(new OnCompleteListener<String>() {
-                    @Override
-                    public void onComplete(@NonNull Task<String> task) {
-                        if (!task.isSuccessful()) {
-                            Log.w("FCM", "Fetching FCM registration token failed", task.getException());
-                            return;
-                        }
-
-                        // Get new FCM registration token
-                        String token = task.getResult();
-                        SessionManager.getInstance().setDeviceToken(token);
-                        // Log and toast
-                        Log.d("FCM", token);
-//                        Toast.makeText(MainActivity.this, token, Toast.LENGTH_SHORT).show();
-                    }
-                });
 
         setFragment(TAG_RALLY, rallyFragment);
